requirements.txt
pandas
beautifulsoup4
optuna
spacy
spacy-cleaner
bertopic
sentence-transformers
scikit-learn
octis
umap-learn
gensim
wordcloud
matplotlib
datetime
spacy-fastlang
corextopic
numpy
scipy
